The project is ready. We still need to figure out how to explicitly delete some objects that I temporarily created.
What's new in this version:
We create objects at the lower level and assign the arguments and/or variables to such object and pass the object to the upper level until we reach a DRCNode where we set the DRCNode name. The temporal objects are called "TempArgNode". Also I put a comment on the function "del" which is the one we should use to delete the temp objects. I'm not sure how to use this function but for now we won't use it.

The comparison is working now. I modified the tokens file and put a new derivation rule

The print function in the drcobject file was updated.